% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlasUtilities.R
\name{get_atlas_concept}
\alias{get_atlas_concept}
\title{Get Concept ID Details}
\usage{
get_atlas_concept(id)
}
\arguments{
\item{id}{Integer representing an OMOP Concept ID}
}
\value{
Properly formatted json text related to the concept ID.
The json text is used for OHDSI Atlas queries
}
\description{
Takes a valid OMOP Concept ID and returns a list of concept fields
}
\examples{
## Retrieves the json text for the concept ID 436665 (bipolar disorder)
bipolar_id <- 436665
get_atlas_concept(bipolar_id)
# {
# "items": [
#   {
#     "concept": {
#       "CONCEPT_ID": 436665,
#       "CONCEPT_NAME": "Bipolar disorder",
#       "STANDARD_CONCEPT": "S",
#       "STANDARD_CONCEPT_CAPTION": "Standard",
#       "INVALID_REASON": "V",
#       "INVALID_REASON_CAPTION": "Valid",
#       "CONCEPT_CODE": "13746004",
#       "DOMAIN_ID": "Condition",
#       "VOCABULARY_ID": "SNOMED",
#       "CONCEPT_CLASS_ID": "Clinical Finding"
#     },
#     "isExcluded": false,
#     "includeDescendants": true,
#     "includeMapped": true
#   }
# ]
# }
}
